fixed space/tab inconsistencies and fixed _i_max datatype inconsistency

+++ Autopilot/Inc/PID.hpp
@@ -18,8 +18,8 @@
 
 class PIDController
 {
-    public:
-    	/**
+	public:
+		/**
 		* Initialises the Pid object.
 		* @param[in]	_kp 	The proportional gain.
 		* @param[in]	_ki 	The integral gain.
@@ -28,9 +28,9 @@ class PIDController
 		* @param[in]	_min_output		The minimum value that can be output, if computations return smaller, the output will be set to this value.
 		* @param[in]	_max_output		The maximum value that can be output, if computations return larger, the output will be set to this value.
 		*/
-        PIDController(float _kp, float _ki, float _kd, float _i_max, float _min_output, float _max_output);
+		PIDController(float _kp, float _ki, float _kd, float _i_max, float _min_output, float _max_output);
 
-    	/**
+		/**
 		* Executes a PID computation.
 		* The PID algorithm uses the derivative of the actual values, rather than the derivatives of the error. That makes it immune to sudden changes in commanded set points.
 		* THIS METHOD MUST BE CALLED AT A REGULAR INTERVAL!!! UNDEFINED BEHAVIOUR WILL OCCUR OTHERWISE.
@@ -40,16 +40,16 @@ class PIDController
 		* @param[in]	actualRate  The current measured derivative (This parameter is optional).
 		* @return					The result of the PID computation.
 		*/
-        float execute(float desired, float actual, float actualRate = std::nanf(""));
+		float execute(float desired, float actual, float actualRate = std::nanf(""));
 
-    private:
+	private:
 
 		float kp, kd, ki;
-		int16_t i_max;
+		float i_max;
 		float integral;
-    	float historicalValue[3];
-    	float min_output;
-    	float max_output;
+		float historicalValue[3];
+		float min_output;
+		float max_output;
 
 };
